Fix clientside routing for <a> tags
Clientside routing with <a> tags did not work when the <a> tag was
wrapped around anything that wasn't pure text, such as a <div>

This happened because the click handler that handled the <a> tag routing
inside router.js would not detect the <a> tag when it was wrapped around
anything that wasn't text.

Thanks to a Stack Overflow answer, which I cited in a comment above the
fix, the click handler is now fixed and actually simplified!